промежуточная версия задания Модуляция до слияния с master
